XRENDERING-563: Handle metadata rendering in HTML macro
  * Define an AbstractHTMLMacroRenderer for all renderers.
