+++ xwiki-rendering-macros/xwiki-rendering-macro-html/src/main/java/org/xwiki/rendering/internal/macro/html/renderers/annotatedxhtml/HTMLMacroAnnotatedXHTMLRenderer.java
@@ -25,16 +25,10 @@ import javax.inject.Named;
 import org.xwiki.component.annotation.Component;
 import org.xwiki.component.annotation.InstantiationStrategy;
 import org.xwiki.component.descriptor.ComponentInstantiationStrategy;
-import org.xwiki.component.phase.Initializable;
-import org.xwiki.component.phase.InitializationException;
-import org.xwiki.rendering.internal.macro.html.HTMLMacroBlockStateChainingListener;
-import org.xwiki.rendering.internal.macro.html.HTMLMacroChainingRenderer;
+import org.xwiki.rendering.internal.macro.html.AbstractHTMLMacroRenderer;
 import org.xwiki.rendering.internal.renderer.xhtml.AnnotatedXHTMLChainingRenderer;
 import org.xwiki.rendering.internal.renderer.xhtml.image.XHTMLImageRenderer;
 import org.xwiki.rendering.internal.renderer.xhtml.link.XHTMLLinkRenderer;
-import org.xwiki.rendering.listener.chaining.EmptyBlockChainingListener;
-import org.xwiki.rendering.listener.chaining.ListenerChain;
-import org.xwiki.rendering.listener.chaining.MetaDataStateChainingListener;
 import org.xwiki.rendering.renderer.AbstractChainingPrintRenderer;
 import org.xwiki.stability.Unstable;
 
@@ -49,7 +43,7 @@ import org.xwiki.stability.Unstable;
 @Named("htmlmacro+annotatedxhtml/1.0")
 @InstantiationStrategy(ComponentInstantiationStrategy.PER_LOOKUP)
 @Unstable
-public class HTMLMacroAnnotatedXHTMLRenderer extends AbstractChainingPrintRenderer implements Initializable
+public class HTMLMacroAnnotatedXHTMLRenderer extends AbstractHTMLMacroRenderer
 {
     /**
      * To render link events into XHTML. This is done so that it's pluggable because link rendering depends on how the
@@ -70,19 +64,8 @@ public class HTMLMacroAnnotatedXHTMLRenderer extends AbstractChainingPrintRender
     private XHTMLImageRenderer imageRenderer;
 
     @Override
-    public void initialize() throws InitializationException
+    protected AbstractChainingPrintRenderer getSyntaxRenderer()
     {
-        ListenerChain chain = new ListenerChain();
-        setListenerChain(chain);
-
-        // Construct the listener chain in the right order. Listeners early in the chain are called before listeners
-        // placed later in the chain.
-        chain.addListener(this);
-        chain.addListener(new HTMLMacroBlockStateChainingListener(chain));
-        chain.addListener(new EmptyBlockChainingListener(chain));
-        chain.addListener(new MetaDataStateChainingListener(chain));
-        chain.addListener(new HTMLMacroChainingRenderer(
-            new AnnotatedXHTMLChainingRenderer(this.linkRenderer, this.imageRenderer, chain))
-        );
+        return new AnnotatedXHTMLChainingRenderer(this.linkRenderer, this.imageRenderer, getListenerChain());
     }
 }
